+++ Tests/HTTPHeaderParserTests/XCTestManifests.swift
@@ -3,7 +3,8 @@ import XCTest
 #if !canImport(ObjectiveC)
 public func allTests() -> [XCTestCaseEntry] {
     return [
-        testCase(HTTPHeaderParserTests.allTests),
+        testCase(LinkTests.allTests),
+        testCase(RetryAfterTests.allTests),
         testCase(UriReferenceTests.allTests)
     ]
 }